requirements.txt
Django~=4.0.3
djangorestframework~=3.13.1
django-cors-headers
twilio
matplotlib~=3.3.3
numpy~=1.18.5
pandas~=1.2.2
Pillow~=9.0.1
PyYAML~=6.0
requests~=2.25.1
seaborn~=0.11.2
torch~=1.10.2+cu113
torchvision~=0.11.3+cu113
tqdm~=4.63.0

opencv-python~=4.5.5.64